Trying to make an endpoint for the "Game" entity Added Game package under server.
Added the unversioned config, repository and service.

+++ server/src/main/java/server/Game/GameRepository.java
@@ -1,7 +1,8 @@
 package server.Game;
 
+import commons.Game;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GameRepository extends JpaRepository <Game, Integer> {
+public interface GameRepository extends JpaRepository <Game, Long> {
 
 }
